experiments/veo-genetic-prompt-optimizer/pyproject.toml
[project]
name = "veo-genetic-prompt-optimizer"
version = "0.1.0"
description = ""
authors = [
    {name = "Chouaieb Nemri",email = "nemri@google.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai (>=1.20.0,<2.0.0)",
    "google-cloud-aiplatform[evaluation] (>=1.97.0,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
